requirements.txt
Pillow
imagehash